requirements.txt
arabic-reshaper==3.0.0
blinker==1.9.0
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
emoji==2.15.0
Flask==3.1.2
fonttools==4.61.1
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.14.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
pyparsing==3.2.5
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.3
Werkzeug==3.1.4
wordcloud==1.9.4
